Remove unneeded backends before we start adding new features + rename
