+++ requirements/development.txt
@@ -2,6 +2,4 @@
 openapi-spec-validator==0.2.9
 pytest==6.0.1
 flask==1.1.2
-falcon==2.0.0
-starlette==0.13.7
 requests==2.24.0
